Rename check-overflow file
